Refactor ERC20 token transfer proxy and it's tests
